CreateRestrictionTool (2)
Issue 9: general approach for  using MapToolCapture is now better understood. Changes have been made to CreateRestrictionTool that now allow geometry to be created. Currently, it not saving correctly, but at least it is there.

An issue to consider in the future is the way the rubber band functions - currently there is no "feedback" for the "next" point.