Introduce a core RestrictionManager module
The TOMsRestrictionManager module will keep track of the current date
and the current proposal.

All dialogs or other objects that work with restrictions will share one
RestrictionManager.

To change the current date or proposal, this will be set on the manager,
the manager will then send out a notification signal to everyone
connected to it, so this can be updated everywhere, throughout the user
interface.

It will also always update the filter string on the map layer.