Working but with issues
Issue is that geometry of each node is used for orientation of icon. To resolve this, need to access original geometry. Not sure how to do this but possibly need to record it in the table so that it is an attribute of the feature.

Also started work on include sign geometries into TOMsGeometry.